+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Compute CP dates after updating related tasks (#855)
 
+### Fixed
+- Disable model in localrep (#848)
+
 ## [0.9.0] 2022-03-01
 
 ### Added